Fixed bug where pollen display would stop showing pollen when a player got down to near zero again.

+++ Assets/Scripts/UI/GameUI.cs
@@ -26,11 +26,11 @@ public class GameUI : MonoBehaviour
 			{
 				if (display.flowerType != null)
 				{
-					int pollen = (int)PlayerInventory.Instance.GetPollen(display.flowerType.Id);
+					float pollen = PlayerInventory.Instance.GetPollen(display.flowerType.Id);
 					if (pollen > 0)
 					{
 						display.Show();
-						display.DisplayPollen(pollen);
+						display.DisplayPollen((int)pollen);
 					}
 					else
 					{
